+++ src/main/java/lotto/Application.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.game.LottoGame;
+
 public class Application {
     public static void main(String[] args) {
         LottoGame lottoGame = new LottoGame();
